+++ modules/executors/inputs/node_inputs.py
@@ -70,12 +70,3 @@ class UserInputNode(GenericExecutor):
 
 
     
-class ConstantNode(ListNode):
-    node_type = "constant"
-    def __init__(self,*args):
-        super().__init__(*args)
-
-    def set_template(self,args):
-        solved_args = try_solve_files(args)
-        packed_args = [solved_args]
-        super().set_template(packed_args)
